+++ wgpu/backend/__init__.py
@@ -1,0 +1,4 @@
+"""
+The backend implementations of the wgpu API.
+You need to select one by importing it, e.g. ``import wgpu.backend.rs``.
+"""
