+++ wgpu/gui/__init__.py
@@ -1,0 +1,4 @@
+"""
+Code to help provide a canvas to render to.
+You don't need this per see, but it makes things a bit easier.
+"""
